Modularization Implementation Example
Implement a 'componentized' architecture in an activity by using a proxy component factory and component container to achieve modular functionality. A component can either be related to a View or unrelated to a View, depending on the requirements. At the same time, components can communicate with each other through a Service, achieving decoupling between components.